requirements.txt

customtkinter

ddgs
googlesearch-python
beautifulsoup4

requests
python-dotenv

phonenumbers

urllib3